+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+## Fixed
+
+- 修正 Chrome 与 Safari 无法正确显示文档的错误；
+
 ## [v6.0.0]
 
 ### Added
@@ -9,7 +13,7 @@
 - 添加 Valid 方法，用于验证文档是否正确；
 - 为 Request 和 Param 添加了五个用于描述 XML 数据的字段；
 - 为 apidoc 元素添加 mimetype，用于指定全局可用的 mimetype 值；
-- 为 Request 和 Param 添加 attr 字段，用于在 content-type 为 xml 的请求中表示属性。
+- 为 Request 和 Param 添加 attr 字段，用于在 content-type 为 xml 的请求中表示属性；
 - 添加 mock 子命令，用于生成 mock 数据的功能；
 - 添加 static 子命令，用于打开查看文档的服务；
 - 添加导出为 openapi 的功能；